+++ Src/WindowsServer/WindowsServer.Shared/AzureWebAppRoleEnvironmentTelemetryInitializer.cs
@@ -15,7 +15,8 @@
     {
         /// <summary>Azure Web App name corresponding to the resource name.</summary>
         private const string WebAppNameEnvironmentVariable = "WEBSITE_SITE_NAME";
-        /// <summary>Azure Web App host that contains site name and slot: sitename-slot.azurewebsites.net.</summary>
+
+        /// <summary>Azure Web App host that contains site name and slot: site-slot.azurewebsites.net.</summary>
         private const string WebAppHostNameEnvironmentVariable = "WEBSITE_HOSTNAME";
 
         private string roleInstanceName;
@@ -56,12 +57,12 @@
 
         private string GetRoleName()
         {
-            return Environment.GetEnvironmentVariable(WebAppNameEnvironmentVariable);
+            return Environment.GetEnvironmentVariable(WebAppNameEnvironmentVariable) ?? string.Empty;
         }
 
         private string GetRoleInstanceName()
         {
-            return Environment.GetEnvironmentVariable(WebAppHostNameEnvironmentVariable);
+            return Environment.GetEnvironmentVariable(WebAppHostNameEnvironmentVariable) ?? string.Empty;
         }
     }
 }